requirements.txt
Brotli==1.1.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cssselect2==0.8.0
fonttools==4.60.1
idna==3.11
iniconfig==2.3.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
pycparser==2.23
pydyf==0.11.0
Pygments==2.19.2
pyphen==0.17.2
pytest==9.0.2
python-dotenv==1.2.1
requests==2.32.5
tinycss2==1.4.0
tinyhtml5==2.0.0
urllib3==2.6.2
weasyprint==66.0
webencodings==0.5.1
zopfli==0.2.3.post1
